move function to .cpp file
